Remove `linkTitle` attribute during an unlink
Given how CKEditor works, meaning that the Attribute Elements are merged
on render to produce an anchor tag with all of the defined attributes,
the unlink command removed only the `linkHref` one, leaving `linkTitle`
in place, which still produces an anchor tag on render. To mitigate that
issue, we remove `linkTitle` as a part of executing the `Unlink`
command.

Trello card:

Co-authored-by: Michał Chudzik <michal.chudzik@futurelearn.com>
Co-authored-by: Tytus Powroźnik <tytus.powroznik@futurelearn.com>